Separate out generic report structure from gbs_prism specifics
